requirements.txt
certifi==2020.6.20
cffi==1.14.2
cryptography==3.1
decorator==4.4.2
dnspython==2.0.0
pycparser==2.20
pymongo==3.11.0
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
XlsxWriter==1.3.4
